+++ satis-service/src/main/java/com/marul/satis/MusteriFeignClient.java
@@ -1,7 +1,6 @@
 package com.marul.satis;
 
 import com.marul.dto.MusteriDto;
-import com.marul.dto.musteri.MusteriResponseDto;
 import com.marul.dto.result.SuccessDataResult;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,5 +10,5 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface MusteriFeignClient {
 
     @GetMapping("/musteri-service/v1/musteri/findById/{musteriId}")
-    SuccessDataResult<MusteriDto> findById(@PathVariable("musteriId") Integer musteriId);
+    SuccessDataResult<MusteriDto> findById(@PathVariable("musteriId") Long musteriId);
 }
